A version that doesn't seem to break
Playing a sequence of moves by repeatedly calling into the wasm module
causes the crash.
Playing the same sequence of moves via Haskell code in a single call to
the wasm module does not cause a crash, and it does not seem possible to
crash after do, even if the game's state is reset.